requirements.txt
flask==3.0.2
requests==2.31.0
dnspython==2.6.1
cryptography==42.0.5
python-whois==0.8.0
beautifulsoup4==4.12.3
python-dotenv==1.0.1
gunicorn==21.2.0
urllib3<2.0.0 